+++ MegaManLofi/IConsoleBuffer.h
@@ -19,13 +19,13 @@ namespace MegaManLofi
       virtual void SetDefaultBackgroundColor( ConsoleColor color ) = 0;
       virtual void Clear() override = 0;
 
-      virtual void Draw( int left, int top, char buffer ) = 0;
-      virtual void Draw( int left, int top, char buffer, ConsoleColor foregroundColor ) = 0;
-      virtual void Draw( int left, int top, char buffer, ConsoleColor foregroundColor, ConsoleColor backgroundColor ) = 0;
-      virtual void Draw( int left, int top, const std::string& buffer ) = 0;
-      virtual void Draw( int left, int top, const std::string& buffer, ConsoleColor foregroundColor ) = 0;
-      virtual void Draw( int left, int top, const std::string& buffer, ConsoleColor foregroundColor, ConsoleColor backgroundColor ) = 0;
-      virtual void Draw( int left, int top, const ConsoleSprite& sprite ) = 0;
+      virtual void Draw( short left, short top, char buffer ) = 0;
+      virtual void Draw( short left, short top, char buffer, ConsoleColor foregroundColor ) = 0;
+      virtual void Draw( short left, short top, char buffer, ConsoleColor foregroundColor, ConsoleColor backgroundColor ) = 0;
+      virtual void Draw( short left, short top, const std::string& buffer ) = 0;
+      virtual void Draw( short left, short top, const std::string& buffer, ConsoleColor foregroundColor ) = 0;
+      virtual void Draw( short left, short top, const std::string& buffer, ConsoleColor foregroundColor, ConsoleColor backgroundColor ) = 0;
+      virtual void Draw( short left, short top, const ConsoleSprite& sprite ) = 0;
 
       virtual void Flip() override = 0;
    };
