+++ MegaManLofiTests/mock_Player.h
@@ -11,10 +11,10 @@ public:
    MOCK_METHOD( bool, IsMoving, ( ), ( const, override ) );
    MOCK_METHOD( const MegaManLofi::Rectangle&, GetHitBox, ( ), ( const, override ) );
    MOCK_METHOD( void, SetDirection, ( MegaManLofi::Direction ), ( override ) );
-   MOCK_METHOD( double, GetVelocityX, ( ), ( const, override ) );
-   MOCK_METHOD( double, GetVelocityY, ( ), ( const, override ) );
-   MOCK_METHOD( void, SetVelocityX, ( double ), ( override ) );
-   MOCK_METHOD( void, SetVelocityY, ( double ), ( override ) );
+   MOCK_METHOD( long long, GetVelocityX, ( ), ( const, override ) );
+   MOCK_METHOD( long long, GetVelocityY, ( ), ( const, override ) );
+   MOCK_METHOD( void, SetVelocityX, ( long long ), ( override ) );
+   MOCK_METHOD( void, SetVelocityY, ( long long ), ( override ) );
    MOCK_METHOD( void, StopX, ( ), ( override ) );
    MOCK_METHOD( void, StopY, ( ), ( override ) );
 };
